Add more parameters for image pre-processing
This includes the ability for the user to choose if they want to grayscale or turn to a 3 channel image, do blurring or not, do hist equalization or not, etc. More freedom in terms of training model as well.

(cherry picked from commit e4a8a26478e6bc086c3fc0628103667c0a2e6d5f)